requirements.txt
pandas
pytest
streamlit
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.21.0